% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ass2Data501.R
\name{calculate_cooks_distance}
\alias{calculate_cooks_distance}
\title{Calculate Cook's Distance
This function calculates Cook's Distance for each observation in a linear model.}
\usage{
calculate_cooks_distance(model, data)
}
\arguments{
\item{model}{A linear model object of class \code{lm}.}

\item{data}{A data frame containing the data used to fit the linear model.
The data should have the same structure and number of rows as the model's data}
}
\value{
A numeric vector of Cook's Distance values for each observation in the model.
}
\description{
Calculate Cook's Distance
This function calculates Cook's Distance for each observation in a linear model.
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + hp, data = mtcars)
cooks_d <- calculate_cooks_distance(model, mtcars)
}
