% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ass2Data501.R
\name{calculate_dffits}
\alias{calculate_dffits}
\title{Calculate DFFITS
This function calculates the DFFITS value for each observation in a linear model.
DFFITS is a measure used to assess the influence of each observation on the fitted values of a regression model.}
\usage{
calculate_dffits(model, data)
}
\arguments{
\item{model}{A linear model object of class \code{lm}.}

\item{data}{A data frame containing the data used to fit the linear model.
The data should have the same structure and number of rows as the model's data}
}
\value{
A numeric vector of DFFITS values for each observation in the model
}
\description{
Calculate DFFITS
This function calculates the DFFITS value for each observation in a linear model.
DFFITS is a measure used to assess the influence of each observation on the fitted values of a regression model.
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + hp, data = mtcars)
dffits_values <- calculate_dffits(model, mtcars)
}
