% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ass2Data501.R
\name{calculate_hadis_influence}
\alias{calculate_hadis_influence}
\title{Calculate Hadi's Influence Measure.}
\usage{
calculate_hadis_influence(model, data)
}
\arguments{
\item{model}{A linear model object of class \code{lm}}

\item{data}{A data frame containing the data used to fit the linear model.}
}
\value{
A numeric vector of Hadi's influence measure values for each observation in the model.
}
\description{
This function calculates Hadi influence measure for each observation in a linear model.
Hadi's influence measure is used to identify influential observations that have an impact on the regression model.
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + hp, data = mtcars)
hadi_values <- calculate_hadis_influence(model, mtcars)
}
