% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ass2Data501.R
\name{plot_influence_measures}
\alias{plot_influence_measures}
\title{Plot Influence Measures for Linear Models
This function plots the influence measures (Cook's Distance, DFFITS, or Hadi's Influence Measure)}
\usage{
plot_influence_measures(model, data, method = "cooks")
}
\arguments{
\item{model}{A linear model object of class \code{lm}.}

\item{data}{A data frame containing the data used to fit the linear model.}

\item{method}{A character string specifying which influence measure to plot.
Must be one of \code{"cooks"}, \code{"dffits"}, or \code{"hadi"}.}
}
\value{
A plot of the selected influence measure with points above the threshold labeled.
}
\description{
Plot Influence Measures for Linear Models
This function plots the influence measures (Cook's Distance, DFFITS, or Hadi's Influence Measure)
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + hp, data = mtcars)
plot_influence_measures (model,mtcars,method="cooks")
}
