% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ass2Data501.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs for Influence Measure Calculations}
\usage{
validate_inputs(data, model)
}
\arguments{
\item{data}{A data frame containing the data used to fit the linear model.}

\item{model}{A linear model object of class \code{lm}.}
}
\value{
No return value since the function is used for checking
}
\description{
This function validates the inputs to ensure they are suitable for calculating influence measures.
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + hp, data = mtcars)
validate_inputs(mtcars, model)
}
